+++ src/features/autolayout/libsbmlnetwork_autolayout_node.cpp
@@ -1,8 +1,8 @@
 #include "libsbmlnetwork_autolayout_node.h"
-#include "../libsbmlnetwork_layout_helpers.h"
-#include "../libsbmlnetwork_common.h"
-#include "../features/user_data/libsbmlnetwork_user_data.h"
-#include "../features/defaults/libsbmlnetwork_defaults_layout.h"
+#include "../../libsbmlnetwork_layout_helpers.h"
+#include "../../libsbmlnetwork_common.h"
+#include "../user_data/libsbmlnetwork_user_data.h"
+#include "../defaults/libsbmlnetwork_defaults_layout.h"
 
 // AutoLayoutNodeBase
